reorg of starter componsnnts
