+++ packages/__cxstarter__/src/components/Header/index.ts
@@ -1,13 +1,1 @@
-import { cxHeader, asHeaderToken } from '@bodiless/cx-layout';
-
-const Default = asHeaderToken({
-  ...cxHeader.Default,
-  Components: {
-    ...cxHeader.Default.Components,
-  },
-});
-
-export const __cxstarter__Header = {
-  ...cxHeader,
-  Default,
-};
+export { __cxstarter__Header } from './tokens';
