+++ packages/__cxstarter__/src/components/Header/tokens/__cxstarter__Header.ts
@@ -1,0 +1,13 @@
+import { cxHeader, asHeaderToken } from '@bodiless/cx-layout';
+
+const Default = asHeaderToken({
+  ...cxHeader.Default,
+  Components: {
+    ...cxHeader.Default.Components,
+  },
+});
+
+export default {
+  ...cxHeader,
+  Default,
+};
