+++ packages/__cxstarter__/src/components/Header/tokens/index.ts
@@ -1,0 +1,3 @@
+import __cxstarter__Header from './__cxstarter__Header';
+
+export { __cxstarter__Header };
