+++ packages/__cxstarter__/src/components/Layout/tokens/index.ts
@@ -1,0 +1,3 @@
+import __cxstarter__Layout from './__cxstarter__Layout';
+
+export { __cxstarter__Layout };
