+++ packages/__cxstarter__/src/index.ts
@@ -1,1 +1,3 @@
 export * from './components/Page';
+export * from './components/Layout';
+export * from './components/Header';
